Augment model unittest
